+++ ShiftsLogger.ManuelE-Osorio/.vscode/.mono/registry/CurrentUser/software/microsoft/csdevkit/telemetry/persistentpropertybag/c57a9efce9b74de382d905a89852db71/values.xml
@@ -1,0 +1,3 @@
+﻿<values>
+  <value name="VS.Core.HardwareId" type="string">931579A8FCE205F7BFFED3B16BF8870F2652435E43DF3C642667A1928A30CE36</value>
+</values>
