+++ ShiftsLogger.ManuelE-Osorio/.vscode/.mono/registry/CurrentUser/software/microsoft/csdevkit/telemetry/persistentpropertybag/values.xml
@@ -1,0 +1,3 @@
+﻿<values>
+  <value name="mac.address" type="string">824541b0dd717897fcb6b2ccc68ea1987c70b908fecd0e6bb59f0b7d1981c167</value>
+</values>
