+++ ShiftsLogger.ManuelE-Osorio/.vscode/.mono/registry/CurrentUser/software/microsoft/csdevkit/v1/pids/16075_133514442938070070/values.xml
@@ -1,0 +1,13 @@
+﻿<values>
+  <value name="StillAlive" type="qword">133514508951516663</value>
+  <value name="extversion" type="string">1.2.7</value>
+  <value name="PID" type="int">16075</value>
+  <value name="VSIs64BitProcess" type="string">True</value>
+  <value name="vscodeversion" type="string">1.83.1</value>
+  <value name="CTime" type="qword">133514442938070070</value>
+  <value name="VSExeVersion" type="string">4.1000.24.5801</value>
+  <value name="ShutdownReason" type="int">3</value>
+  <value name="VSInstanceId" type="string">b8f16c6b-8a25-4659-80bd-0b3b635ef9501706970692368</value>
+  <value name="extname" type="string">ms-dotnettools.csdevkit</value>
+  <value name="ExeName" type="string">Microsoft.CodeAnalysis.LanguageServer</value>
+</values>
