+++ ShiftsLogger.ManuelE-Osorio/.vscode/.mono/registry/CurrentUser/software/microsoft/sqmclient/values.xml
@@ -1,0 +1,4 @@
+﻿<values>
+  <value name="UserId" type="string">{CABF476E-0721-4193-A9FE-965B922DBFDB}</value>
+  <value name="MachineId" type="string">{824541B0-DD71-7897-FCB6-B2CCC68EA198}</value>
+</values>
